@@ -6,16 +6,16 @@
     "start": "node server"
   },
   "dependencies": {
-    "apisauce": "^0.1.3",
+    "apisauce": "^0.3.0",
     "babel-polyfill": "^6.7.4",
     "ramda": "^0.21.0",
     "ramdasauce": "^1.0.0",
-    "react": "^0.14.8",
-    "react-dom": "^0.14.8",
+    "react": "^15.2.1",
+    "react-dom": "^15.2.1",
     "react-redux": "^4.4.5",
     "redux": "^3.4.0",
     "redux-logger": "^2.6.1",
-    "redux-saga": "^0.9.5",
+    "redux-saga": "^0.11.0",
     "seamless-immutable": "^6.0.0",
     "socket.io": "^1.4.5",
     "socket.io-client": "^1.4.5"
