@@ -1,9 +1,8 @@
 import React, { Component, PropTypes } from 'react'
 import { connect } from 'react-redux'
 import Actions from './Actions/Creators'
-import Reactotron from '../client'
 
-export default class RootContainer extends Component {
+class RootContainer extends Component {
 
   constructor (props) {
     super(props)
@@ -13,23 +12,22 @@ export default class RootContainer extends Component {
   handlePress (e) {
     e.preventDefault()
     const {dispatch} = this.props
-    Reactotron.log('A touchable was pressed.')
-    dispatch(Actions.requestTemperature('Toronto'))
+    console.tron.log('A touchable was pressed. 🦄')
+    dispatch(Actions.requestGithub())
   }
 
-  componentWillMount () {
+  componentDidMount () {
     const { dispatch } = this.props
     dispatch(Actions.startup())
   }
 
   render () {
-    const {city, temperature, fetching} = this.props
+    const {message} = this.props
     return (
       <div style={Styles.container}>
         <a onClick={this.handlePress}>
-          <p style={Styles.weather}>
-            {`The weather in ${city} is ${fetching ? 'loading' : temperature}.`}
-          </p>
+          <h3>Last Commit Message</h3>
+          <p style={Styles.message}>{message}</p>
         </a>
       </div>
     )
@@ -44,7 +42,7 @@ const Styles = {
     justifyContent: 'center',
     alignItems: 'center'
   },
-  weather: {
+  message: {
     fontFamily: 'sans-serif',
     fontSize: '25px',
     textAlign: 'center'
@@ -53,16 +51,12 @@ const Styles = {
 
 RootContainer.propTypes = {
   dispatch: PropTypes.func,
-  city: PropTypes.string,
-  temperature: PropTypes.number,
-  fetching: PropTypes.bool
+  message: PropTypes.string
 }
 
 const mapStateToProps = (state) => {
   return {
-    city: state.weather.city,
-    temperature: state.weather.temperature,
-    fetching: state.weather.fetching
+    message: state.github.message
   }
 }
 
