@@ -1,6 +1,6 @@
 export default {
   STARTUP: 'STARTUP',
-  TEMPERATURE_REQUEST: 'TEMPERATURE_REQUEST',
-  TEMPERATURE_RECEIVE: 'TEMPERATURE_RECEIVE',
-  TEMPERATURE_FAILURE: 'TEMPERATURE_FAILURE'
+  GITHUB_REQUEST: 'GITHUB_REQUEST',
+  GITHUB_RECEIVE: 'GITHUB_RECEIVE',
+  GITHUB_FAILURE: 'GITHUB_FAILURE'
 }
