@@ -1,5 +1,4 @@
 import React, { Component } from 'react'
-import Layout from './Layout'
 import Root from './Root'
 import { Provider } from 'react-redux'
 import configureStore from './Store/Store'
@@ -10,9 +9,7 @@ export default class App extends Component {
   render () {
     return (
       <Provider store={store}>
-        <Layout>
-          <Root />
-        </Layout>
+        <Root />
       </Provider>
     )
   }
